Add 'cancel' class to back buttons to deactivate jQuery.validate.
jQuery.validate checks for a 'cancel' class on the triggering button and
skips validation if it’s found. All buttons to previous steps should
have this class.

The alternative is to upgrade jQuery.validate which is bundled in
clientside_validation. The fix can be removed when we drop support for
such old versions